Split files by class
