+++ Xamarin.Forms.Platform.UAP/Xamarin.Forms.Platform.UAP.csproj
@@ -42,6 +42,7 @@
     <Compile Include="AccessibilityExtensions.cs" />
     <Compile Include="CollectionView\ItemsViewRenderer.cs" />
     <Compile Include="CollectionView\SelectableItemsViewRenderer.cs" />
+    <Compile Include="CollectionView\StructuredItemsViewRenderer.cs" />
     <Compile Include="ColorExtensions.cs" />
     <Compile Include="DispatcherProvider.cs" />
     <Compile Include="Extensions\ImageExtensions.cs" />
